refactor: align directory layout with kei-lisp conventions
- App.vue を <Repl/> をマウントするだけの薄いルートに分離し、REPL ロジックは src/components/Repl/Repl.vue に抽出
- src/composables/useReplOutput.ts を kei-lisp と同様の per-folder 構成（useReplOutput/index.ts）に変更
- src/shims/ 配下を役割が一目で分かる命名にリネーム
  - process.ts        → installProcessShim.ts
  - v8.ts / vm.ts / module.ts → nodeV8.ts / nodeVm.ts / nodeModule.ts
- 上記に伴い vite.config.ts の alias パスと main.ts の import を更新
