+++ vite.config.ts
@@ -20,12 +20,12 @@ export default defineConfig({
   resolve: {
     alias: {
       // kei-lisp imports node:v8 for the (gc) builtin; provide a no-op stub.
-      'node:v8': fileURLToPath(new URL('./src/shims/v8.ts', import.meta.url)),
+      'node:v8': fileURLToPath(new URL('./src/shims/nodeV8.ts', import.meta.url)),
       // kei-lisp also imports node:vm for the same (gc) builtin; throwing stub.
-      'node:vm': fileURLToPath(new URL('./src/shims/vm.ts', import.meta.url)),
+      'node:vm': fileURLToPath(new URL('./src/shims/nodeVm.ts', import.meta.url)),
       // kei-lisp's Repl uses createRequire from node:module. We never run Repl
       // in the browser, so a throwing stub satisfies bundling.
-      'node:module': fileURLToPath(new URL('./src/shims/module.ts', import.meta.url)),
+      'node:module': fileURLToPath(new URL('./src/shims/nodeModule.ts', import.meta.url)),
     },
   },
 });
